add useDotLottieButton hook
